+++ src/primaite/simulator/system/services/database/database_service.py
@@ -1,5 +1,6 @@
 from ipaddress import IPv4Address
 from typing import Any, Dict, List, Literal, Optional, Union
+from uuid import uuid4
 
 from primaite import getLogger
 from primaite.simulator.file_system.file_system import File
@@ -57,7 +58,7 @@ class DatabaseService(Service):
 
         # check if the backup server was configured
         if self.backup_server_ip is None:
-            self.sys_log.error(f"{self.name} - {self.sys_log.hostname}: not configured.")
+            self.sys_log.warning(f"{self.name} - {self.sys_log.hostname}: not configured.")
             return False
 
         software_manager: SoftwareManager = self.software_manager
@@ -110,7 +111,7 @@ class DatabaseService(Service):
         db_file = self.file_system.get_file(folder_name="database", file_name="database.db", include_deleted=True)
 
         if db_file is None:
-            self.sys_log.error("Database file not initialised.")
+            self.sys_log.warning("Database file not initialised.")
             return False
 
         # if the file was deleted, get the old visible health state
@@ -145,8 +146,16 @@ class DatabaseService(Service):
         """Returns the database folder."""
         return self.file_system.get_folder_by_id(self.db_file.folder_id)
 
+    def _generate_connection_id(self) -> str:
+        """Generate a unique connection ID."""
+        return str(uuid4())
+
     def _process_connect(
-        self, connection_id: str, password: Optional[str] = None
+        self,
+        src_ip: IPv4Address,
+        connection_request_id: str,
+        password: Optional[str] = None,
+        session_id: Optional[str] = None,
     ) -> Dict[str, Union[int, Dict[str, bool]]]:
         """Process an incoming connection request.
 
@@ -158,24 +167,24 @@ class DatabaseService(Service):
         :rtype: Dict[str, Union[int, Dict[str, bool]]]
         """
         status_code = 500  # Default internal server error
+        connection_id = None
         if self.operating_state == ServiceOperatingState.RUNNING:
             status_code = 503  # service unavailable
             if self.health_state_actual == SoftwareHealthState.OVERWHELMED:
-                self.sys_log.error(
-                    f"{self.name}: Connect request for {connection_id=} declined. Service is at capacity."
-                )
+                self.sys_log.error(f"{self.name}: Connect request for {src_ip=} declined. Service is at capacity.")
             if self.health_state_actual == SoftwareHealthState.GOOD:
                 if self.password == password:
                     status_code = 200  # ok
+                    connection_id = self._generate_connection_id()
                     # try to create connection
-                    if not self.add_connection(connection_id=connection_id):
+                    if not self.add_connection(connection_id=connection_id, session_id=session_id):
                         status_code = 500
-                        self.sys_log.info(f"{self.name}: Connect request for {connection_id=} declined")
+                        self.sys_log.warning(f"{self.name}: Connect request for {connection_id=} declined")
                     else:
                         self.sys_log.info(f"{self.name}: Connect request for {connection_id=} authorised")
                 else:
                     status_code = 401  # Unauthorised
-                    self.sys_log.info(f"{self.name}: Connect request for {connection_id=} declined")
+                    self.sys_log.warning(f"{self.name}: Connect request for {connection_id=} declined")
         else:
             status_code = 404  # service not found
         return {
@@ -183,6 +192,7 @@ class DatabaseService(Service):
             "type": "connect_response",
             "response": status_code == 200,
             "connection_id": connection_id,
+            "connection_request_id": connection_request_id,
         }
 
     def _process_sql(
@@ -206,7 +216,7 @@ class DatabaseService(Service):
         self.sys_log.info(f"{self.name}: Running {query}")
 
         if not self.db_file:
-            self.sys_log.info(f"{self.name}: Failed to run {query} because the database file is missing.")
+            self.sys_log.error(f"{self.name}: Failed to run {query} because the database file is missing.")
             return {"status_code": 404, "type": "sql", "data": False}
 
         if query == "SELECT":
@@ -276,7 +286,7 @@ class DatabaseService(Service):
                 return {"status_code": 401, "data": False}
         else:
             # Invalid query
-            self.sys_log.info(f"{self.name}: Invalid {query}")
+            self.sys_log.warning(f"{self.name}: Invalid {query}")
             return {"status_code": 500, "data": False}
 
     def describe_state(self) -> Dict:
@@ -299,19 +309,34 @@ class DatabaseService(Service):
         :return: True if the Status Code is 200, otherwise False.
         """
         result = {"status_code": 500, "data": []}
-
         # if server service is down, return error
         if not self._can_perform_action():
             return False
 
         if isinstance(payload, dict) and payload.get("type"):
             if payload["type"] == "connect_request":
+                src_ip = kwargs.get("frame").ip.src_ip_address
                 result = self._process_connect(
-                    connection_id=payload.get("connection_id"), password=payload.get("password")
+                    src_ip=src_ip,
+                    password=payload.get("password"),
+                    connection_request_id=payload.get("connection_request_id"),
+                    session_id=session_id,
                 )
             elif payload["type"] == "disconnect":
                 if payload["connection_id"] in self.connections:
-                    self.remove_connection(connection_id=payload["connection_id"])
+                    connection_id = payload["connection_id"]
+                    connected_ip_address = self.connections[connection_id]["ip_address"]
+                    frame = kwargs.get("frame")
+                    if connected_ip_address == frame.ip.src_ip_address:
+                        self.sys_log.info(
+                            f"{self.name}: Received disconnect command for {connection_id=} from {connected_ip_address}"
+                        )
+                        self.terminate_connection(connection_id=payload["connection_id"], send_disconnect=False)
+                    else:
+                        self.sys_log.warning(
+                            f"{self.name}: Ignoring disconnect command for {connection_id=} as the command source "
+                            f"({frame.ip.src_ip_address}) doesn't match the connection source ({connected_ip_address})"
+                        )
             elif payload["type"] == "sql":
                 if payload.get("connection_id") in self.connections:
                     result = self._process_sql(